ANDROID: Makefile: append BUILD_NUMBER to version string when defined
In current implementation the Linux version string that is attached to
the kernel linux image as well as the all kernel modules includes the
sha1 of the main git project for the kernel.

This Commit adds the BUILD_NUMBER env variable with an 'ab' prefix to
the Linux version string. This maps to both the Linux Image and all
kernel modules version strings.

Bug: 159842160
Bug: 137521202
Test: Build the kernel and test the version string
Change-Id: I4795955f36f477b257ca86495726e24d6b196d05
Signed-off-by: Alex Hong <rurumihong@google.com>
(cherry picked from commit e60f995403c9799dd0dc14d62d77d6698a8a7126)
[willmcvicker: fix uts_len check to include -ab$(BUILD_NUMBER)]
Signed-off-by: Will McVicker <willmcvicker@google.com>
Signed-off-by: atndko <z1281552865@gmail.com>