Revert "scsi: ufs-qcom: Add flag to keep track of PHY's power state"
This reverts commit 173e5a10b22ee5a1ce3941f48bbb8320e9768d27.

This can contribute a race condition between is_phy_pwr_on and phy_power_on/off.
Instead of it, we must rely on mutex with power_count in phy_power_on/off more
precisely. And, in order to fix the origina issue caused by multiple power on/
off, we must use power_count as a flag.

So, the final approach is same as moving "is_phy_pwr_on" under mutex.

Bug: 139262967
Change-Id: If71a7d6b8adc316895a3e038047d3e3d860e213e
Signed-off-by: Jaegeuk Kim <jaegeuk@google.com>
Signed-off-by: atndko <z1281552865@gmail.com>