cnss2: set ERR_NOTIFY bit when target enters RDDM
CNSS_DEV_ERR_NOTIFY is not set when rddm is supported.
And if target goes to RDDM during wlan driver probe
routine, device will be shut down. This makes FW dump
not able to be collected since PCIe link is already
suspended.

Fix is to set CNSS_DEV_ERR_NOTIFY bit if system goes
to RDDM state when driver is in loading state.

Change-Id: I2153ddfc632f88e234ef0192fb7a25d0f71bc98d
Signed-off-by: jiad <jiad@codeaurora.org>