clk: qcom: osm: avoid freq run over max_freq
Setting max freq which is not in the given cpu's freq table, the
target_freq is the same with max freq. When selecting the freq, it uses
CPUFREQ_RELATION_L, so the frequency round to the next. This is
reasonable for normal cases but the frequency should not run over the
max setting.

Bug: 158134249
Test: build and set max/min freq to check
Change-Id: I804454225c10b7a023b323c58e77b7bba1ec3576
Signed-off-by: Kyle Lin <kylelin@google.com>
Signed-off-by: atndko <z1281552865@gmail.com>