msm: camera: cci: Fix incorrect use of cci config ioctl
The cci configuration will be transitioned to a new API that does not
require routing through the v4l layer. This is work-in-progrss so in
the mean time prevent the device from being exposed as configurable
from userspace.
The ioctl will still be exposed to kernel users so fix the arg size
as well. We want size of struct not pointer.

Bug: 170138526
Test: build pass, local verify camera function

Change-Id: I1eeb58b8355a81fc36b9845e2096060b03f1debe
Signed-off-by: Om Parkash <oparkash@codeaurora.org>
Signed-off-by: atndko <z1281552865@gmail.com>