drivers: thermal: notify at least one thermal zone if no trips are violated
During any sensor threshold notification, the of-thermal framework
checks whether any trip of each thermal zone for that sensor is
violated or not. It requests thermal zone re-evaluation only for
those thermal zones those trip is violated. If trips are violated
for none of thermal zones due to some reason, then of-thermal ignores
that sensor notification and it leads to no active trip threshold
for that sensor forever.

Notify and request thermal zone re-evaluation for at least one thermal
zone for that sensor if trips are violated for none in of-thermal.
It makes sure that for every sensor notification, thermal zone
aggregate and set next active trip.

Change-Id: If9955977fe32929d853815ddb9408385291b3c58
Signed-off-by: Manaf Meethalavalappu Pallikunhi <manafm@codeaurora.org>
Signed-off-by: atndko <z1281552865@gmail.com>