crypto: msm: compat_qcedev: Fix invalid memory access
In a system when userspace is 32 bit and kernel is 64,
the 32 bit pointers when passed to kernel, this pointer
value is assigned to internal kernel space pointer which
is 64 bit but the upper 32 bit of kernel pointer may hold
garbage or invalid data.

To ensure the kernel pointer holds valid value, first
assign NULL to it, to clear entire 64 bit and then assign
userspace pointer value to it.

Change-Id: I3b4f287c9b18129d1265956118b1806f80cff5ba
Signed-off-by: Jiten Patel <jitepate@codeaurora.org>
Signed-off-by: atndko <z1281552865@gmail.com>