drm/msm/sde: Wait for ctl_start interrupt asynchronously for cmd mode panels
Waiting for ctl_start when completing commits on command mode panels
usually takes around 5 ms even though in most cases the next commit will
arrive after long after that. This constitutes an unnecessary delay in
the display commit pipeline.

To remedy this, wait for the ctl_start interrupt asynchronously. When a
new commit comes in before the ctl_start interrupt fires off for the
previous commit, the new commit will be stalled until the ctl_start
interrupt fires off. This is achieved using a kthread worker that runs
at the same priority as other display-commit kthreads, so as not to
stall the display commit pipeline by waiting for something running at a
low priority.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: atndko <z1281552865@gmail.com>