net: usbnet: Resubmit URBs from ISR for interfaces without throttling
'Commit 39e4df8d18bb ("net: usbnet: Add mechanism to throttle
usb0 RX traffic in USB SS")' added regression that the completed
URBs are not getting resubmitted to the HW for interfaces which
do not require dynamic throttling or when the USB2 link is used.
This leads to poor throughput on such interfaces. Fix this by
adding proper checks.

Change-Id: I9d127068f40a30a6ea900ab98ab3518f8b789ae3
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>