usb: phy: Run PHY POR sequence as a part of bus resume
Currently the driver is just enabling clocks as a part of bus
resume. This leads to PHY not being in a proper state on exit
from XO shutdown. The observed issue is that the PHY is stuck in
resume signal to the connected peripheral for a long time,
eventually leading to disconnect-connect.
Fix this by running Power On Reset sequence in bus resume.

Change-Id: I100688e70fa5be03fe5bc11bbd49f07b3c5d1b76
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>