ARM: dts: msm: Expand access of DWC registers for sdmshrike
The multiport controller on sdmshrike is a host-only core. The
DSTS(Device Status) register does not exist in the register space
and hence cannot be read to get the link state to e.g. ensure if
the link is in compliance mode or not. The GDBGLTSSM register can
be read for that. Hence expand the DWC register space access to
cover the same.

Change-Id: I9bd46c9bf909e1cacd8f0a69a836563076341e52
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>