serial: msm_geni_serial: Enable core interrupt for hsuart
Enablement of core interrupt will happen in runtime_resume
for hsuart, but with auto-suspend-disable flag runtime resume/suspend
will not invoke which further leads to not enable the interrupts. Due
to this the data transfers are getting failed on both tx and rx paths.

Fix this by enable the core interrupt while port_open and
disable it during port_close.

Change-Id: I92cfe6edfc4329b48afdf00de40298fb08ec59b3
Signed-off-by: venkata manasa kakarla <venkka@codeaurora.org>