scsi: ufs: serialise ufs probe for multiple ufs instance
For the platform having multiple ufs instance,we always
have 1st instance as the boot device. Due to async
scan of the device, there can be scenarios where
2nd ufs instance gets probed first and get enumerated.
Later when user space try to mount rootfs on that partition
we get permission error.

To handle this we are deferring the probe of 2nd ufs instance
until the 1st UFS is probed and LUNs are detected.

This will not impact boot kpi for the 1st instance
which is boot device but may impact boot kpi
for the 2nd instance.

Change-Id: I7a2f543ff30eca68370e975f9271d9518eeabd48
Signed-off-by: Nitin Rawat <nitirawa@codeaurora.org>
Signed-off-by: Hiren Gohel <vhiren@codeaurora.org>