Serial: msm_geni_serial: Remove the vote in GSI mode for TX
Currently for GSI mode, we are not removing the vote at the end
of stop tx sequencer and due to this if we cancel an on-going Tx
transfer then there is vote imbalance because we put the vote at
start of TX transfer and removing a vote at the end of TX transfer.
Due to the vote being active, LPM test will fail. To avoid it,
remove the vote.

Change-Id: I8b5d621bf0cde5b9f3eb579d6ffa70be0603f4d9
Signed-off-by: Aniket Randive <arandive@codeaurora.org>