sched: tune: be more conservative about boosting top app
Currently with top-app boost value being fixed to 1 we get a tonn of
random small tasks hanging on the big cluster that have frankly no business
staying there. Let's be smarter about top-app group and only allow
apps that are actually on the display (w/ adj == 0) to stay on big cores.

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: atndko <z1281552865@gmail.com>