fbdev: msm: handle the div value reset by mdss driver
The pclk div value will set in pixel_div_set_div 
function and save in the reg DSIPHY_SSC9. But this 
value will be reset to 0 after mdss_dsi_phy_disable 
and regulator_disable(mdata->fs), this will lead to 
the correct div value to default value 0. So we store 
the correct value and fill it back when get div.

Change-Id: I8eb4f08f1320e9dc410247b12cb1912d6c7613f0
Signed-off-by: longzhao <quic_longzhao@quicinc.com>