extcon: Allocate blocking_notifier_head after registering the device
Commit 6e721f3ad0535b24f19a62420f4da95212cf069c
("extcon: Modify extcon device to be created after driver data is set")
move device_register right before adding extcon_dev_list.

This is fine on upstream, however since we have blocking notifier and
it uses managed kzalloc (which has to be used when the dev is already
registered successfully), we have to move it as well.

Fixes: a4febd54c189fc2cb63205461bdd58e909e0fb66
("extcon: Add usage of blocking notifier chain")
Signed-off-by: Nauval Rizky <enuma.alrizky@gmail.com>
Signed-off-by: atndko <z1281552865@gmail.com>